backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# OpenClaw Agent Framework
openclaw>=0.1.0

# WaveSpeed AI SDK
wavespeed>=0.1.0
openai>=1.40.0

# Canvas LMS
canvasapi==3.2.0

# Vector Store & RAG
chromadb==0.5.3
sentence-transformers==3.0.1
rank-bm25==0.2.2
llama-index-core==0.10.57

# Web Scraping
scrapy==2.11.1
beautifulsoup4==4.12.3
lxml==5.2.2

# PDF Processing
pymupdf==1.24.5
pdfplumber==0.11.0

# Browser Automation
playwright==1.44.0

# Google APIs
google-api-python-client==2.130.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8

# --- 核心修改区：解决依赖冲突 ---
# Config (放宽限制，迎合 LangChain 等下游库的要求)
pydantic-settings>=2.10.1,<3.0.0
python-dotenv==1.0.1

# 隐式依赖的护栏 (防止底层的 Protobuf 和 AWS 库打架)
protobuf>=3.20.2,<5.0.0
boto3<1.34.70

# Scheduling & Email
apscheduler==3.10.4
aiosmtplib==3.0.1

# Utils
httpx==0.27.0
aiofiles==23.2.1icalendar>=5.0
